requirements/dev.txt
django-environ
django-filter
django-allauth
djangorestframework
djangorestframework-simplejwt
djangorestframework-camel-case
drf-flex-fields
drf-aggregation
dj-rest-auth
drf-yasg
drf-ip-restrictions
markdown
psycopg2-binary
urllib3==1.26.6
django-cors-headers
casbin
django-cleanup
sorl-thumbnail-serializer-field==0.2.1
sorl-thumbnail
Pillow
django-storages[google]==1.13.1
faker